Remove useless file.
