+++ packages/core/src/common/services/entity-serializer.service.spec.ts
@@ -1,9 +1,12 @@
 import * as chai from 'chai';
 import * as chaiAsPromised from 'chai-as-promised';
 import {
+  Column,
   createConnection,
+  Entity,
   getConnection,
   getManager,
+  PrimaryGeneratedColumn,
 } from 'typeorm';
 
 import { ObjectDoesNotExist } from '../errors';
@@ -13,8 +16,25 @@ chai.use(chaiAsPromised);
 const expect = chai.expect;
 
 import { EntitySerializer } from './entity-serializer.service';
-// TODO: remove user.entity.spec file.
-import { User } from './user.entity.spec';
+
+@Entity()
+export class User {
+  @PrimaryGeneratedColumn()
+  // @ts-ignore : Property 'id' has no initializer and is not definitely assigned in theconstructor.
+  id: number;
+
+  @Column()
+  // @ts-ignore : Property 'firstName' has no initializer and is not definitely assigned in theconstructor.
+  firstName: string;
+
+  @Column()
+  // @ts-ignore : Property 'lastName' has no initializer and is not definitely assigned in theconstructor.
+  lastName: string;
+
+  @Column({ default: false })
+  // @ts-ignore : Property 'isAdmin' has no initializer and is not definitely assigned in theconstructor.
+  isAdmin: boolean;
+}
 
 function testSuite(type: 'mysql'|'mariadb'|'postgres'|'sqlite', connectionName: string) {
